% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/materials.R
\name{materials}
\alias{materials}
\title{Broswe Japanese useR Community's material}
\usage{
materials(community = c("HijiyamaR", "Hiroshima.R", "Kashiwa.R", "Nagoya.R",
  "Osaka.R", "SappoRo.R", "Tokyo.R", "Tsukuba.R", "Yokohama.R"),
  number = NULL, session = NULL, browse = FALSE)
}
\arguments{
\item{community}{a character string. select useR community for one of the strings "HijiyamaR", "Hiroshima.R", "Kashiwa.R", "Nagoya.R", "Osaka.R", "SappoRo.R", "Tokyo.R", "Tsukuba.R" or "Yokohama.R".}

\item{number}{}

\item{session}{}

\item{browse}{a logical value indicading wheater the browse material.}
}
\description{
Broswe Japanese useR Community's material
}
\examples{
materials(community = "Tokyo.R")
}

