% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/speaker.R
\name{speaker}
\alias{speaker}
\title{Listup favorite spearker's material}
\usage{
speaker(id = NULL)
}
\arguments{
\item{id}{user name}
}
\description{
Listup favorite spearker's material
}
\examples{
speaker(id = "u_ribo")
}

